requirements.txt
yfpy
pyyaml
python-dotenv
